requirements.txt
fastapi[standard]
sqlalchemy 
pwdlib
pyjwt
psycopg2-binary
streamlit
requests